# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Portfolio Dashboard API Documentation
  description: Complete API documentation for Portfolio Dashboard with Admin & Public authentication, OTP verification, and user management
  version: 2.0.0
  contact:
    name: Md Afzal Ansari
    email: mdafzal14777@gmail.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
paths:
...
  /admin-auth/register:
    post:
      summary: Register New Admin
      description: Register a new admin user with OTP verification
      tags:
        - Admin Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - user_name
                - email
                - password
              properties:
                name:
                  type: string
                  example: "Admin User"
                  minLength: 2
                  maxLength: 50
                user_name:
                  type: string
                  example: "adminuser"
                  minLength: 3
                  maxLength: 20
                email:
                  type: string
                  format: email
                  example: "admin@example.com"
                password:
                  type: string
                  example: "password123"
                  minLength: 8
      responses:
        "200":
          description: Admin registration successful - OTP sent to email
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Admin registration successful! Please check your email to verify your account."
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      email:
                        type: string
                      userType:
                        type: string
                        example: "admin"
                      expiresIn:
                        type: number
                        example: 600
        "400":
          description: Validation error
        "409":
          description: User already exists
...
  /admin-auth/login:
    post:
      summary: Admin Login Request
      description: Initiate admin login with OTP verification
      tags:
        - Admin Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: OTP sent to email for login verification
...
  /public-auth/register:
    post:
      summary: Register New Public User
      description: Register a new public user with OTP verification
      tags:
        - Public Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - user_name
                - email
                - password
              properties:
                name:
                  type: string
                  example: "Public User"
                user_name:
                  type: string
                  example: "publicuser"
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  example: "password123"
      responses:
        "200":
          description: Public user registration successful - OTP sent to email
...
  /public-auth/login:
    post:
      summary: Public User Login Request
      description: Initiate public user login with OTP verification
      tags:
        - Public Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: OTP sent to email for login verification
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# Tags
...